# static/api-docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "LeetCode Scraper API",
    "description": "API for fetching LeetCode problems and user statistics",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
  "paths": {
    "/scrape-leetcode": {
      "get": {
        "summary": "Get LeetCode problems",
        "description": "Retrieves a paginated list of LeetCode problems with optional filtering",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Number of problems per page",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search query to filter problems by name",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "difficulty",
            "in": "query",
            "description": "Filter problems by difficulty level",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["all", "Easy", "Medium", "Hard"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "problems": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "example": "1"
                          },
                          "title": {
                            "type": "string",
                            "example": "Two Sum"
                          },
                          "difficulty": {
                            "type": "string",
                            "enum": ["Easy", "Medium", "Hard"]
                          },
                          "acceptance_rate": {
                            "type": "number",
                            "format": "float",
                            "example": 48.5
                          },
                          "url": {
                            "type": "string",
                            "example": "https://leetcode.com/problems/two-sum/"
                          }
                        }
                      }
                    },
                    "total": {
                      "type": "integer",
                      "example": 2000
                    },
                    "current_page": {
                      "type": "integer",
                      "example": 1
                    },
                    "total_pages": {
                      "type": "integer",
                      "example": 40
                    },
                    "per_page": {
                      "type": "integer",
                      "example": 50
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user-stats": {
      "post": {
        "summary": "Get user statistics",
        "description": "Retrieves statistics for a specific LeetCode user",
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "LeetCode username"
                  }
                },
                "required": ["username"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "profile": {
                          "type": "object",
                          "properties": {
                            "ranking": {
                              "type": "integer",
                              "example": 100000
                            },
                            "reputation": {
                              "type": "integer",
                              "example": 100
                            },
                            "starRating": {
                              "type": "number",
                              "format": "float",
                              "example": 4.5
                            }
                          }
                        },
                        "submitStats": {
                          "type": "object",
                          "properties": {
                            "acSubmissionNum": {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "difficulty": {
                                    "type": "string",
                                    "enum": ["EASY", "MEDIUM", "HARD"]
                                  },
                                  "count": {
                                    "type": "integer"
                                  }
                                }
                              }
                            }
                          }
                        },
                        "totalProblems": {
                          "type": "object",
                          "properties": {
                            "total": {
                              "type": "integer",
                              "example": 2000
                            },
                            "byDifficulty": {
                              "type": "object",
                              "properties": {
                                "Easy": {
                                  "type": "integer",
                                  "example": 500
                                },
                                "Medium": {
                                  "type": "integer",
                                  "example": 1000
                                },
                                "Hard": {
                                  "type": "integer",
                                  "example": 500
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Username is required"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/problem-counts": {
      "get": {
        "summary": "Get problem counts",
        "description": "Retrieves the total count of problems and counts by difficulty",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer",
                          "example": 2000
                        },
                        "byDifficulty": {
                          "type": "object",
                          "properties": {
                            "Easy": {
                              "type": "integer",
                              "example": 500
                            },
                            "Medium": {
                              "type": "integer",
                              "example": 1000
                            },
                            "Hard": {
                              "type": "integer",
                              "example": 500
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}